requirements.txt
chainlit
langchain
langchain-community
langchain-openai
openai
sentence-transformers
pymupdf
faiss-cpu
python-dotenv
langchain-huggingface
aiofiles